refactor(sessions): extract List composer
- Create List.tsx with session list integration logic
  (Active/Recent split, recentGroupRef, handleRemoveSession wrapper,
  scroll container)
- Create List.test.tsx with 9 integration tests covering list
  rendering, session clicking, removal, and scroll behavior
- Update Sidebar.tsx to use <List /> component,
  removing session-related logic
- Migrate session list tests from Sidebar.test.tsx to List.test.tsx
  (kept 8 sidebar chrome tests)
- Tests: npm run test passing (1741 tests)
